Create MoveWithJoystickCommand class plus minor changes to OI class.
For control via a single joystick, there is now the
MoveWithJoystickCommand class. In order to implement this class, a
RobotDrive object was created in the DriveTrain subsystem, allowing
control of the robot to be as simple as calling a method and passing the
joystick as a parameter. The OI class was also modified. The buttons are
now all contained in a private HashMap field, and are accessed via the
getButton(String key) method. Joystick is now also private and is
accessed by the getJoystick() method.